Abstracted some table components, abstracted different parts of user profile page to be cleaner, removed places where setters were passed directly into React functions, allow for dynamic pageSize in some components, updated React hooks and components to reflect that new requirement
