+++ js/src/app/user/[userId]/_components/UserProfile/UserProfileHeaderSkeleton.tsx
@@ -1,0 +1,5 @@
+import { Skeleton } from "@mantine/core";
+
+export default function UserProfileHeaderSkeleton() {
+  return <Skeleton height={32} width={200} mx={"md"} mb={"xs"} />;
+}
