+++ js/src/app/user/[userId]/_components/UserSubmissions/hooks.ts
@@ -1,7 +1,6 @@
 import { useURLState } from "@/lib/hooks/useUrlState";
 import { ApiResponse } from "@/lib/types/apiResponse";
 import { Question } from "@/lib/types/db/question";
-import { User } from "@/lib/types/db/user";
 import { Page } from "@/lib/types/page";
 import { keepPreviousData, useQuery } from "@tanstack/react-query";
 import { useCallback, useEffect } from "react";
@@ -10,10 +9,12 @@ export const useUserSubmissionsQuery = ({
   userId,
   initialPage = 1,
   tieToUrl = false,
+  pageSize = 20,
 }: {
   userId?: string;
   initialPage?: number;
   tieToUrl?: boolean;
+  pageSize?: number;
 }) => {
   const [page, setPage] = useURLState("page", initialPage, tieToUrl);
   const [searchQuery, setSearchQuery, debouncedQuery] = useURLState(
@@ -46,7 +47,7 @@ export const useUserSubmissionsQuery = ({
   const query = useQuery({
     queryKey: ["submission", "user", userId, page, debouncedQuery],
     queryFn: () =>
-      fetchUserSubmissions({ page, userId, query: debouncedQuery }),
+      fetchUserSubmissions({ page, userId, query: debouncedQuery, pageSize }),
     placeholderData: keepPreviousData,
   });
 
@@ -66,29 +67,18 @@ async function fetchUserSubmissions({
   page,
   userId,
   query,
+  pageSize,
 }: {
   page: number;
   userId?: string;
   query?: string;
+  pageSize: number;
 }) {
   const response = await fetch(
-    `/api/user/${userId ?? ""}/submissions?page=${page}&query=${query}`,
+    `/api/user/${userId ?? ""}/submissions?page=${page}&query=${query}&pageSize=${pageSize}`,
   );
 
-  const json = (await response.json()) as ApiResponse<
-    Page<User & { questions: Question[] }>
-  >;
+  const json = (await response.json()) as ApiResponse<Page<Question[]>>;
 
-  if (json.success) {
-    return { ...json.data, success: json.success, message: json.message };
-  }
-
-  return {
-    hasNextPage: false,
-    data: null,
-    pages: 0,
-    message: json.message,
-    success: json.success,
-    pageSize: 0,
-  };
+  return json;
 }
