+++ js/src/lib/router.tsx
@@ -5,7 +5,7 @@ import LoginPage from "@/app/login/Login.page";
 import Onboarding from "@/app/onboarding/Onboarding.page";
 import RootPage from "@/app/Root.page";
 import SubmissionDetailsPage from "@/app/submission/s/[submissionId]/SubmissionDetails.page";
-import UserSubmissionsPage from "@/app/user/[userId]/UserSubmissions.page";
+import UserSubmissionsPage from "@/app/user/[userId]/UserProfile.page";
 import { createBrowserRouter } from "react-router-dom";
 
 export const router = createBrowserRouter([
